setting pause duration works
