+++ styles.js
@@ -147,4 +147,15 @@ ul {
   padding-inline-start: 0;
 
 }
+
+/* hide arrows on number inputs */
+input[type=number]::-webkit-outer-spin-button,
+input[type=number]::-webkit-inner-spin-button {
+  -webkit-appearance: none; /* Chrome, Safari, Edge */
+  margin: 0;
+}
+
+input[type=number] {
+  -moz-appearance: textfield; /* Firefox */
+}
 `;
